requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.0
openai
pandas==2.2.0
phidata 
agno 
langchain